# docs/contracts/v1/openapi.json
{
  "components": {
    "schemas": {
...
      "DevelopmentLoginRequest": {
        "properties": {
          "username": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username"
        ],
        "title": "DevelopmentLoginRequest",
        "type": "object"
      },
      "DevelopmentLoginResponse": {
        "properties": {
          "access_token": {
            "title": "Access Token",
            "type": "string"
          },
          "expires_at": {
            "format": "date-time",
            "title": "Expires At",
            "type": "string"
          },
          "principal": {
            "$ref": "#/components/schemas/PrincipalView"
          },
          "token_type": {
            "const": "Bearer",
            "default": "Bearer",
            "title": "Token Type",
            "type": "string"
          },
          "warning": {
            "default": "Development authentication: this token was issued without verifying any credential and is valid only on this process.",
            "title": "Warning",
            "type": "string"
          }
        },
        "required": [
          "access_token",
          "expires_at",
          "principal"
        ],
        "title": "DevelopmentLoginResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PrincipalView": {
        "description": "The caller, as the API sees them.",
        "properties": {
          "capabilities": {
            "description": "Exactly what this principal may do. The frontend reads this to decide what to show; the backend does not trust that decision.",
            "items": {
              "type": "string"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "development": {
            "title": "Development",
            "type": "boolean"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "expires_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "inactive_roles": {
            "description": "Reserved roles this account holds that grant nothing yet.",
            "items": {
              "type": "string"
            },
            "title": "Inactive Roles",
            "type": "array"
          },
          "roles": {
            "items": {
              "type": "string"
            },
            "title": "Roles",
            "type": "array"
          },
          "source": {
            "enum": [
              "oidc",
              "development",
              "collector_key"
            ],
            "title": "Source",
            "type": "string"
          },
          "subject": {
            "title": "Subject",
            "type": "string"
          },
          "unrecognized_roles": {
            "description": "Role or group values in the token that ADG does not know. Shown so an administrator can see a misconfigured assignment instead of guessing.",
            "items": {
              "type": "string"
            },
            "title": "Unrecognized Roles",
            "type": "array"
          }
        },
        "required": [
          "subject",
          "display_name",
          "email",
          "source",
          "development",
          "roles",
          "capabilities"
        ],
        "title": "PrincipalView",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "Windows-domain share-access auditing and governance",
    "title": "ADG API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/auth/dev/login": {
      "post": {
        "description": "Issue a token for a configured development account.\n\nRegistered only when ``ADG_AUTH_MODE=development``; see :func:`build_auth_router`.",
        "operationId": "development_login_auth_dev_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DevelopmentLoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DevelopmentLoginResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "No such development account."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "DEVELOPMENT ONLY: sign in as a configured account without a credential",
        "tags": [
          "auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "me_auth_me_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrincipalView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "The authenticated caller",
        "tags": [
          "auth"
        ]
      }
    },
...
  }
}